+++ src/boot_O2/loadfragment2.c
@@ -71,7 +71,7 @@ void Load2_Relocate(u32 allocatedVRamAddr, OverlayRelocationSection* overlayInfo
 s32 Load2_LoadOverlay(u32 vRomStart, u32 vRomEnd, u32 vRamStart, u32 vRamEnd, u32 allocatedVRamAddr) {
     int nbytes;
     u32 pad;
-    u32 size;
+    size_t size;
     void* end;
     OverlayRelocationSection* overlayInfo;
 
@@ -108,7 +108,7 @@ s32 Load2_LoadOverlay(u32 vRomStart, u32 vRomEnd, u32 vRamStart, u32 vRamEnd, u3
 
 void* Overlay_AllocateAndLoad(u32 vRomStart, u32 vRomEnd, u32 vRamStart, u32 vRamEnd) {
     void* allocatedVRamAddr;
-    u32 size;
+    size_t size;
 
     size = vRamEnd - vRamStart;
     allocatedVRamAddr = StartHeap_AllocR(size);
